Product update

+++ src/main/java/by/tms/onlinerclone26onl/controller/CatalogController.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.util.Base64;
 import java.util.List;
 import java.util.Map;
@@ -27,8 +28,8 @@ public class CatalogController {
     @GetMapping("/{idProduct}") //TODO доработать ссылку
     public String getProduct(@PathVariable("idProduct") long idProduct, Model model) {
         Product product = productService.getProductById(idProduct);
-        List<Double> pricesMinAndMax = productService.searchPriceMinAndMax(idProduct);
-        Map<Long, Double> allPrice = productService.searchAllPrice(idProduct);
+        List<BigDecimal> pricesMinAndMax = productService.searchPriceMinAndMax(idProduct);
+        Map<Long, BigDecimal> allPrice = productService.searchAllPrice(idProduct);
         model.addAttribute("allPrice", allPrice);
         String photo = Base64.getEncoder().encodeToString(product.getPhoto());
         model.addAttribute("photo", photo);
